New test for priority
