+++ array.pbs
@@ -1,0 +1,16 @@
+#!/usr/bin/env /bin/bash
+
+#PBS -N arrayJob
+#PBS -o arrayJob_%A_%a.out
+#PBS -t 1-5
+#PBS -l walltime=01:00:00
+#PBS -l naccelerators=1
+#PBS -l proc=gpu6gb
+
+######################
+# Begin work section #
+######################
+
+echo "My SLURM_ARRAY_JOB_ID:" $SLURM_ARRAY_JOB_ID
+echo "My SLURM_ARRAY_TASK_ID: " $SLURM_ARRAY_TASK_ID
+nvidia-smi
